Move comm metadata modal to page, first pass
