+++ client/scripts/views/pages/manage_community/admin_panel_tabs.ts
