+++ client/scripts/views/pages/manage_community/chain_metadata_management_table.ts
