+++ client/scripts/views/pages/manage_community/community_metadata_management_table.ts
