+++ client/styles/pages/manage_community.scss
@@ -1,6 +1,6 @@
 @import 'client/styles/shared';
 
-.ManageCommunityModal {
+.ManageCommunityPage {
     width: 800px;
     @include sm-max {
         width: initial;
